refactor(domain): 카드 생성 시, 셔플 전략을 사용하도록 수정
