+++ src/main/java/domain/deck/CardShuffleStrategy.java
@@ -1,0 +1,9 @@
+package domain.deck;
+
+import domain.card.Card;
+
+import java.util.List;
+
+public interface CardShuffleStrategy {
+    void shuffle(List<Card> cards);
+}
