+++ src/main/java/domain/deck/RandomShuffleStrategy.java
@@ -1,0 +1,13 @@
+package domain.deck;
+
+import domain.card.Card;
+
+import java.util.Collections;
+import java.util.List;
+
+public class RandomShuffleStrategy implements CardShuffleStrategy {
+    @Override
+    public void shuffle(List<Card> cards) {
+        Collections.shuffle(cards);
+    }
+}
