+++ src/main/java/domain/game/GameResult.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import domain.participant.Dealer;
 import domain.participant.Player;
