+++ src/main/java/domain/game/WinningStatus.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import domain.participant.Dealer;
 import domain.participant.Player;
