+++ src/test/java/domain/DeckTest.java
@@ -1,9 +1,13 @@
 package domain;
 
 import domain.card.Card;
+import domain.deck.CardShuffleStrategy;
+import domain.deck.Deck;
+import domain.deck.RandomShuffleStrategy;
 import org.junit.jupiter.api.Test;
 
-import java.util.List;
+import java.util.HashSet;
+import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -11,17 +15,23 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 class DeckTest {
 
     @Test
-    void 중복되지_않은_52장의_카드_묶음을_만든다() {
+    void 중복되지_않은_52장의_카드_덱을_생성한다() {
+        CardShuffleStrategy cardShuffleStrategy = new RandomShuffleStrategy();
+        Deck deck = Deck.createDeck(cardShuffleStrategy);
 
-        List<Card> cards = Deck.createDeck();
-        Deck deck = new Deck(cards);
         assertThat(deck.size()).isEqualTo(52);
+
+        Set<Card> cards = new HashSet<>();
+        while (deck.size() > 0) {
+            cards.add(deck.draw());
+        }
+
+        assertThat(cards.size()).isEqualTo(52);
     }
 
     @Test
     void 덱이_비었을_때_예외를_발생한다() {
-        List<Card> cards = Deck.createDeck();
-        Deck deck = new Deck(cards);
+        Deck deck = Deck.createDeck(new RandomShuffleStrategy());
         int deckSize = deck.size();
 
         for (int i = 0; i < deckSize; i++) {
@@ -34,8 +44,7 @@ class DeckTest {
 
     @Test
     void 덱에서_카드를_한_장_뽑으면_덱의_사이즈가_감소한다() {
-        List<Card> cards = Deck.createDeck();
-        Deck deck = new Deck(cards);
+        Deck deck = Deck.createDeck(new RandomShuffleStrategy());
         int size = deck.size();
 
         deck.draw();
