+++ src/test/java/domain/HandTest.java
@@ -3,6 +3,7 @@ package domain;
 import domain.card.Card;
 import domain.card.Rank;
 import domain.card.Suit;
+import domain.participant.Hand;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
