+++ src/test/java/domain/NameTest.java
@@ -1,9 +1,9 @@
 package domain;
 
+import domain.participant.Name;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 class NameTest {
