+++ src/test/java/domain/ParticipantTest.java
@@ -1,6 +1,8 @@
 package domain;
 
 import domain.card.Card;
+import domain.deck.Deck;
+import domain.deck.RandomShuffleStrategy;
 import domain.participant.Participant;
 import domain.participant.Player;
 import org.junit.jupiter.api.BeforeEach;
@@ -22,8 +24,7 @@ class ParticipantTest {
 
     @Test
     void 참가자들은_카드를_받는다() {
-        List<Card> cards = Deck.createDeck();
-        Deck deck = new Deck(cards);
+        Deck deck = Deck.createDeck(new RandomShuffleStrategy());
         pobi.receive(deck.draw());
 
         assertThat(pobi.handSize()).isEqualTo(startSize + 1);
@@ -31,8 +32,7 @@ class ParticipantTest {
 
     @Test
     void 참가자들은_시작_시_카드_두장을_받는다() {
-        List<Card> cards = Deck.createDeck();
-        Deck deck = new Deck(cards);
+        Deck deck = Deck.createDeck(new RandomShuffleStrategy());
         pobi.receiveInitialCards(deck.drawInitialCards());
 
         assertThat(pobi.handSize()).isEqualTo(startSize + 2);
